Add Javascript implementation of task, to run online
Add Javascript implementation of high level construal task so it
can be run online at pavlovia.org. The Javascript implementation has
custom code, specifically the loading of resources in psychoJS.start()
and how actions and scenarios are randomly selected in
scenario_trialsLoopBegin() and action_trialsLoopBegin() that will be
overwritten if the python implementation is used to auto-generate
the Javascript implementation.

Also, action_conditions.csv has a dummy column, to workaround an issue
loading a csv conditions file with just one column.